refactor(GiniCaptureSDK): Improve QR code parsing readability
- Split per-format QR test suites into dedicated QRCodeFormats files
- Rename LZMADecoder members (p, rc) to probs, rangeDecoder
- Expand format switch to one case per line and add clarifying comments

PP-2324
