+++ CaptureSDK/GiniCaptureSDK/Tests/GiniCaptureSDKTests/GiniQRCodeDocumentTests.swift
@@ -485,314 +485,3 @@ final class GiniQRCodeDocumentTests: XCTestCase {
                                                                    withConfig: giniConfiguration))
     }
 }
-
-// MARK: - SPC (Swiss Payment Code / QR-bill)
-
-@Suite("SPC QR Code")
-struct SPCQRCodeTests {
-
-    private func makeSPC(referenceType: String = "QRR",
-                         reference: String = "210000000003139471430009017",
-                         iban: String = "DE89370400440532013000") -> String {
-        var lines = Array(repeating: "", count: 31)
-        lines[0]  = "SPC"
-        lines[1]  = "0200"
-        lines[2]  = "1"
-        lines[3]  = iban
-        lines[4]  = "S"
-        lines[5]  = "Test Recipient"
-        lines[18] = "1234.50"
-        lines[19] = "CHF"
-        lines[27] = referenceType
-        lines[28] = reference
-        return lines.joined(separator: "\n")
-    }
-
-    @Test func detectedAsSPC() {
-        let doc = GiniQRCodeDocument(scannedString: makeSPC())
-        #expect(doc.qrCodeFormat == .spc)
-    }
-
-    @Test func extractsIBANRecipientAndAmount() {
-        let doc = GiniQRCodeDocument(scannedString: makeSPC())
-        #expect(doc.extractedParameters["iban"] == "DE89370400440532013000")
-        #expect(doc.extractedParameters["paymentRecipient"] == "Test Recipient")
-        #expect(doc.extractedParameters["amountToPay"] == "1234.50:CHF")
-    }
-
-    @Test func extractsReferenceWhenQRR() {
-        let doc = GiniQRCodeDocument(scannedString: makeSPC())
-        #expect(doc.extractedParameters["paymentReference"] == "210000000003139471430009017")
-    }
-
-    @Test func noReferenceWhenNON() {
-        let doc = GiniQRCodeDocument(scannedString: makeSPC(referenceType: "NON",
-                                                            reference: "210000000003139471430009017"))
-        #expect(doc.extractedParameters["paymentReference"] == nil)
-    }
-
-    @Test func nilFormatForInvalidIBAN() {
-        let doc = GiniQRCodeDocument(scannedString: makeSPC(iban: "NOTANIBAN"))
-        #expect(doc.qrCodeFormat == nil)
-    }
-}
-
-// MARK: - SPD (Czech/Slovak Payment Descriptor)
-
-@Suite("SPD QR Code")
-struct SPDQRCodeTests {
-
-    private let validSPD = "SPD*1.0*ACC:CZ6508000000192000145399*AM:100.50*CC:CZK*RN:Test Recipient*MSG:Invoice 123*"
-
-    @Test func detectedAsSPD() {
-        let doc = GiniQRCodeDocument(scannedString: validSPD)
-        #expect(doc.qrCodeFormat == .spd)
-    }
-
-    @Test func extractsAllFields() {
-        let doc = GiniQRCodeDocument(scannedString: validSPD)
-        #expect(doc.extractedParameters["iban"] == "CZ6508000000192000145399")
-        #expect(doc.extractedParameters["amountToPay"] == "100.50:CZK")
-        #expect(doc.extractedParameters["paymentRecipient"] == "Test Recipient")
-        #expect(doc.extractedParameters["paymentReference"] == "Invoice 123")
-    }
-
-    @Test func amountHasNoCurrencySuffixWhenCCMissing() {
-        let doc = GiniQRCodeDocument(scannedString: "SPD*1.0*ACC:CZ6508000000192000145399*AM:100.50*RN:Test Recipient*")
-        #expect(doc.extractedParameters["amountToPay"] == "100.50")
-    }
-
-    @Test func ibanNilWhenACCMissing() {
-        let doc = GiniQRCodeDocument(scannedString: "SPD*1.0*AM:100.50*CC:CZK*RN:Test Recipient*")
-        #expect(doc.extractedParameters["iban"] == nil)
-    }
-
-    @Test func combinesAmountAndCurrencyWhenCCPrecedesAM() {
-        // AM and CC may appear in any order; currency must not be dropped when CC comes first.
-        let doc = GiniQRCodeDocument(scannedString: "SPD*1.0*ACC:CZ6508000000192000145399*CC:CZK*AM:100.50*")
-        #expect(doc.extractedParameters["amountToPay"] == "100.50:CZK")
-    }
-}
-
-// MARK: - Pay by Square (Slovak compressed QR)
-
-@Suite("Pay by Square QR Code")
-struct PayBySquareQRCodeTests {
-
-    let config = GiniConfiguration()
-
-    // 16-char base32hex string. Upper nibble of bytes[0] = bysquareType != 0, so decode() returns nil.
-    private let detectionString = "ABCDEF0123456789"
-
-    @Test func detectedAsPayBySquare() {
-        let doc = GiniQRCodeDocument(scannedString: detectionString)
-        #expect(doc.qrCodeFormat == .payBySquare)
-    }
-
-    @Test func nonBase32HexStringNotDetected() {
-        // 'W' is outside the base32hex alphabet (0–9, A–V); uppercasing won't rescue it
-        let doc = GiniQRCodeDocument(scannedString: "WWWWWWWWWWWWWWWW")
-        #expect(doc.qrCodeFormat != .payBySquare)
-    }
-
-    @Test func corruptPayloadYieldsEmptyParameters() {
-        let doc = GiniQRCodeDocument(scannedString: detectionString)
-        #expect(doc.extractedParameters.isEmpty)
-    }
-
-    @Test func corruptPayloadFailsValidation() {
-        let doc = GiniQRCodeDocument(scannedString: detectionString)
-        #expect(throws: DocumentValidationError.qrCodeFormatNotValid) {
-            try GiniCaptureDocumentValidator.validate(doc, withConfig: config)
-        }
-    }
-
-    // Official bysquare test vector from the bysquare spec (Confluence page 1300430853).
-    // Encodes: IBAN SK2811000000002620154106, amount 12 EUR, variable symbol 2017001,
-    // payment note "PAY bysquare - platba za webovú službu balík Wise".
-    private let officialTestVector =
-        "0008400060RT11GHI1H3ICS8RR40PJJKAMLODMK50MI251UDD11UNM7E306OLN8AGMJUTE" +
-        "SOV4TTES0CMS44EP8OJ9JO3RQP89UE7GME4GHQ9GG62L461V517BI4186SI8J5KT45VUGH" +
-        "OOG9AM35MC87I22BUPU8O2HQLVCDV1DMSQOT1BMEGH00"
-
-    @Test func officialTestVectorDetectedAsPayBySquare() {
-        let doc = GiniQRCodeDocument(scannedString: officialTestVector)
-        #expect(doc.qrCodeFormat == .payBySquare)
-    }
-
-    @Test func officialTestVectorDecodesIBAN() {
-        let doc = GiniQRCodeDocument(scannedString: officialTestVector)
-        #expect(doc.extractedParameters["iban"] == "SK2811000000002620154106")
-    }
-
-    @Test func officialTestVectorDecodesAmount() {
-        let doc = GiniQRCodeDocument(scannedString: officialTestVector)
-        #expect(doc.extractedParameters["amountToPay"] == "12:EUR")
-    }
-
-    @Test func officialTestVectorDecodesPaymentReference() {
-        // variableSymbol "2017001" + " " + paymentNote "PAY bysquare - platba za webovú službu balík Wise"
-        let expected = "2017001 PAY bysquare - platba za webov\u{00FA} službu bal\u{00ED}k Wise"
-        let doc = GiniQRCodeDocument(scannedString: officialTestVector)
-        #expect(doc.extractedParameters["paymentReference"] == expected)
-    }
-
-    @Test func officialTestVectorPassesValidation() {
-        let doc = GiniQRCodeDocument(scannedString: officialTestVector)
-        #expect(throws: Never.self) {
-            try GiniCaptureDocumentValidator.validate(doc, withConfig: config)
-        }
-    }
-}
-
-// MARK: - LZMA Decoder
-
-@Suite("LZMA Decoder")
-struct LZMADecoderTests {
-
-    // Plaintext = a 256-byte block repeated twice. The second block is a match at
-    // distance 256, which forces posSlot >= 14 and therefore exercises
-    // `decodeDirectBits`. Regression guard for the bit-inversion bug that silently
-    // corrupted any decode with a back-reference distance >= 128 bytes.
-    private static func expectedPlaintext() -> [UInt8] {
-        let block = (0..<256).map { UInt8(($0 * 7 + 3) & 0xFF) }
-        return block + block
-    }
-
-    // Raw LZMA1 range-coder stream (bysquare framing: first byte 0x00, bytes 1–4
-    // are the initial `code`) produced by an LZMA1 encoder with lc=3, lp=0, pb=2.
-    private static let lzmaLongDistanceStream: [UInt8] = [
-        0x00, 0x01, 0x82, 0xE0, 0x56, 0x89, 0xC2, 0xF1, 0xF0, 0xE5, 0xDF, 0xF5, 0xE1, 0xD0, 0x31, 0xAE,
-        0x11, 0x00, 0x5A, 0x08, 0x8A, 0x0E, 0x25, 0x1B, 0xF1, 0x5E, 0x0E, 0x99, 0x56, 0x73, 0x92, 0x16,
-        0xE5, 0x5D, 0xC5, 0x02, 0x56, 0xA1, 0xA0, 0x01, 0x75, 0xBF, 0x0A, 0xF7, 0x39, 0xCA, 0x31, 0xCF,
-        0xEE, 0xB4, 0x90, 0xDD, 0x60, 0xD7, 0xD6, 0x99, 0x78, 0xE0, 0x0E, 0x01, 0xC6, 0xAE, 0x95, 0xD4,
-        0x6F, 0xEB, 0x3C, 0x70, 0x99, 0x8F, 0xE6, 0x7C, 0xF8, 0x04, 0x3C, 0xF7, 0x01, 0x32, 0xF5, 0xD7,
-        0x4C, 0xE8, 0x62, 0x21, 0x64, 0x3C, 0x9B, 0x63, 0xA9, 0x15, 0xCA, 0xEB, 0x12, 0x33, 0x35, 0x8E,
-        0x65, 0x45, 0xAD, 0xD6, 0xEF, 0x02, 0x17, 0x75, 0xAC, 0x52, 0x8F, 0xCD, 0x17, 0x57, 0xD1, 0x8A,
-        0x38, 0x71, 0x59, 0xAA, 0xD1, 0x95, 0x74, 0x63, 0x4C, 0x72, 0x7F, 0x8C, 0x52, 0xBC, 0x00, 0x45,
-        0x00, 0xB2, 0xAF, 0xDD, 0xAA, 0x4F, 0xC0, 0x7F, 0x95, 0xFF, 0x78, 0xCB, 0xF9, 0xB1, 0x0F, 0x95,
-        0x27, 0x3C, 0x4F, 0x21, 0x8B, 0xC6, 0xBD, 0x95, 0x41, 0x4E, 0x97, 0x45, 0x34, 0xDE, 0xDD, 0x4A,
-        0x08, 0xE8, 0xC7, 0x0F, 0x84, 0x7C, 0x33, 0xF0, 0x13, 0x69, 0xA4, 0x80, 0x64, 0xB3, 0xFB, 0x73,
-        0x41, 0x6B, 0x11, 0xFF, 0xDC, 0x97, 0x95, 0x3F, 0xC2, 0x34, 0x98, 0x79, 0xA7, 0xA1, 0xAE, 0xAB,
-        0xFC, 0x34, 0xCB, 0xCD, 0xDF, 0xEE, 0x48, 0x13, 0xB8, 0x2E, 0xAE, 0xAE, 0x39, 0xED, 0xFF, 0xC5,
-        0xC0, 0x8F, 0x08, 0xC8, 0x90, 0x77, 0x30, 0x83, 0x27, 0xEB, 0x0A, 0x34, 0xC7, 0x5B, 0x5A, 0x2A,
-        0xF4, 0x3B, 0x2B, 0xEB, 0x87, 0x57, 0x73, 0x33, 0xFF, 0xEC, 0xBE, 0xD4, 0x62, 0xAE, 0xFC, 0xC5,
-        0xF9, 0xFC, 0x08, 0x91, 0x9D, 0xE2, 0x43, 0xFF, 0xFF, 0xD1, 0x9B, 0x62, 0xE0,
-    ]
-
-    @Test func decodesLongDistanceMatchCorrectly() {
-        let expected = Self.expectedPlaintext()
-        let decoded = LZMADecoder.decode(input: Self.lzmaLongDistanceStream,
-                                         outputLength: expected.count)
-        #expect(decoded == expected)
-    }
-
-    @Test func decodedSecondBlockMatchesFirst() {
-        // The bug corrupted only the copied (matched) region; assert both halves
-        // are byte-identical so a future regression can't pass on the literals alone.
-        let expected = Self.expectedPlaintext()
-        let decoded = LZMADecoder.decode(input: Self.lzmaLongDistanceStream,
-                                         outputLength: expected.count)
-        #expect(decoded?.prefix(256) == decoded?.suffix(256))
-    }
-}
-
-// MARK: - UPNQR (Slovenian UPN QR)
-
-@Suite("UPNQR QR Code")
-struct UPNQRQRCodeTests {
-
-    private func makeUPNQR(amountCents: String = "00000048050",
-                            reference: String = "SI00 1234-5678",
-                            iban: String = "DE89370400440532013000",
-                            bic: String = "LJBASI2X",
-                            payee: String = "Gini d.o.o.") -> String {
-        var lines = Array(repeating: "", count: 20)
-        lines[0]  = "UPNQR"
-        lines[8]  = amountCents
-        lines[12] = reference
-        lines[13] = bic
-        lines[14] = iban
-        lines[16] = payee
-        return lines.joined(separator: "\n")
-    }
-
-    @Test func detectedAsUPNQR() {
-        let doc = GiniQRCodeDocument(scannedString: makeUPNQR())
-        #expect(doc.qrCodeFormat == .upnqr)
-    }
-
-    @Test func extractsAllFields() {
-        let doc = GiniQRCodeDocument(scannedString: makeUPNQR())
-        #expect(doc.extractedParameters["iban"] == "DE89370400440532013000")
-        #expect(doc.extractedParameters["paymentRecipient"] == "Gini d.o.o.")
-        #expect(doc.extractedParameters["amountToPay"] == "480.50:EUR")
-        #expect(doc.extractedParameters["bic"] == "LJBASI2X")
-        #expect(doc.extractedParameters["paymentReference"] == "SI00 1234-5678")
-    }
-
-    @Test func convertsCentsToDecimalEUR() {
-        let doc = GiniQRCodeDocument(scannedString: makeUPNQR(amountCents: "00000048050"))
-        #expect(doc.extractedParameters["amountToPay"] == "480.50:EUR")
-    }
-
-    @Test func emptyReferenceOmitsPaymentReference() {
-        let doc = GiniQRCodeDocument(scannedString: makeUPNQR(reference: ""))
-        #expect(doc.extractedParameters["paymentReference"] == nil)
-    }
-
-    @Test func fallsBackToPayerReferenceWhenPaymentReferenceIsEmpty() {
-        var lines = Array(repeating: "", count: 20)
-        lines[0]  = "UPNQR"
-        lines[4]  = "SI01-999"
-        lines[8]  = "0000010000"
-        lines[14] = "DE89370400440532013000"
-        lines[16] = "Gini d.o.o."
-        let doc = GiniQRCodeDocument(scannedString: lines.joined(separator: "\n"))
-        #expect(doc.extractedParameters["paymentReference"] == "SI01-999")
-    }
-}
-
-// MARK: - HUB3 (Croatian PDF417)
-
-@Suite("HUB3 QR Code")
-struct HUB3QRCodeTests {
-
-    private func makeHUB3(currency: String = "EUR",
-                           amountCents: String = "000000000010000",
-                           payee: String = "Gini d.o.o.",
-                           iban: String = "DE89370400440532013000",
-                           reference: String = "HR00 1234567890") -> String {
-        var lines = Array(repeating: "", count: 14)
-        lines[0]  = "HRVHUB30"
-        lines[1]  = currency
-        lines[2]  = amountCents
-        lines[6]  = payee
-        lines[9]  = iban
-        lines[11] = reference
-        return lines.joined(separator: "\n")
-    }
-
-    @Test func detectedAsHUB3() {
-        let doc = GiniQRCodeDocument(scannedString: makeHUB3())
-        #expect(doc.qrCodeFormat == .hub3)
-    }
-
-    @Test func extractsAllFields() {
-        let doc = GiniQRCodeDocument(scannedString: makeHUB3())
-        #expect(doc.extractedParameters["iban"] == "DE89370400440532013000")
-        #expect(doc.extractedParameters["paymentRecipient"] == "Gini d.o.o.")
-        #expect(doc.extractedParameters["amountToPay"] == "100.00:EUR")
-        #expect(doc.extractedParameters["paymentReference"] == "HR00 1234567890")
-    }
-
-    @Test func convertsCentsToDecimalEUR() {
-        let doc = GiniQRCodeDocument(scannedString: makeHUB3(amountCents: "000000000010000"))
-        #expect(doc.extractedParameters["amountToPay"] == "100.00:EUR")
-    }
-
-    @Test func paymentReferenceUsesCallNumberFieldDirectly() {
-        let doc = GiniQRCodeDocument(scannedString: makeHUB3(reference: "HR99 555-666"))
-        #expect(doc.extractedParameters["paymentReference"] == "HR99 555-666")
-    }
-}
