+++ CaptureSDK/GiniCaptureSDK/Tests/GiniCaptureSDKTests/QRCodeFormats/LZMADecoderTests.swift
@@ -1,0 +1,59 @@
+//
+//  LZMADecoderTests.swift
+//
+//  Copyright © 2026 Gini GmbH. All rights reserved.
+//
+
+import Testing
+import XCTest
+@testable import GiniCaptureSDK
+
+@Suite("LZMA Decoder")
+struct LZMADecoderTests {
+
+    // Plaintext = a 256-byte block repeated twice. The second block is a match at
+    // distance 256, which forces posSlot >= 14 and therefore exercises
+    // `decodeDirectBits`. Regression guard for the bit-inversion bug that silently
+    // corrupted any decode with a back-reference distance >= 128 bytes.
+    private static func expectedPlaintext() -> [UInt8] {
+        let block = (0..<256).map { UInt8(($0 * 7 + 3) & 0xFF) }
+        return block + block
+    }
+
+    // Raw LZMA1 range-coder stream (bysquare framing: first byte 0x00, bytes 1–4
+    // are the initial `code`) produced by an LZMA1 encoder with lc=3, lp=0, pb=2.
+    private static let lzmaLongDistanceStream: [UInt8] = [
+        0x00, 0x01, 0x82, 0xE0, 0x56, 0x89, 0xC2, 0xF1, 0xF0, 0xE5, 0xDF, 0xF5, 0xE1, 0xD0, 0x31, 0xAE,
+        0x11, 0x00, 0x5A, 0x08, 0x8A, 0x0E, 0x25, 0x1B, 0xF1, 0x5E, 0x0E, 0x99, 0x56, 0x73, 0x92, 0x16,
+        0xE5, 0x5D, 0xC5, 0x02, 0x56, 0xA1, 0xA0, 0x01, 0x75, 0xBF, 0x0A, 0xF7, 0x39, 0xCA, 0x31, 0xCF,
+        0xEE, 0xB4, 0x90, 0xDD, 0x60, 0xD7, 0xD6, 0x99, 0x78, 0xE0, 0x0E, 0x01, 0xC6, 0xAE, 0x95, 0xD4,
+        0x6F, 0xEB, 0x3C, 0x70, 0x99, 0x8F, 0xE6, 0x7C, 0xF8, 0x04, 0x3C, 0xF7, 0x01, 0x32, 0xF5, 0xD7,
+        0x4C, 0xE8, 0x62, 0x21, 0x64, 0x3C, 0x9B, 0x63, 0xA9, 0x15, 0xCA, 0xEB, 0x12, 0x33, 0x35, 0x8E,
+        0x65, 0x45, 0xAD, 0xD6, 0xEF, 0x02, 0x17, 0x75, 0xAC, 0x52, 0x8F, 0xCD, 0x17, 0x57, 0xD1, 0x8A,
+        0x38, 0x71, 0x59, 0xAA, 0xD1, 0x95, 0x74, 0x63, 0x4C, 0x72, 0x7F, 0x8C, 0x52, 0xBC, 0x00, 0x45,
+        0x00, 0xB2, 0xAF, 0xDD, 0xAA, 0x4F, 0xC0, 0x7F, 0x95, 0xFF, 0x78, 0xCB, 0xF9, 0xB1, 0x0F, 0x95,
+        0x27, 0x3C, 0x4F, 0x21, 0x8B, 0xC6, 0xBD, 0x95, 0x41, 0x4E, 0x97, 0x45, 0x34, 0xDE, 0xDD, 0x4A,
+        0x08, 0xE8, 0xC7, 0x0F, 0x84, 0x7C, 0x33, 0xF0, 0x13, 0x69, 0xA4, 0x80, 0x64, 0xB3, 0xFB, 0x73,
+        0x41, 0x6B, 0x11, 0xFF, 0xDC, 0x97, 0x95, 0x3F, 0xC2, 0x34, 0x98, 0x79, 0xA7, 0xA1, 0xAE, 0xAB,
+        0xFC, 0x34, 0xCB, 0xCD, 0xDF, 0xEE, 0x48, 0x13, 0xB8, 0x2E, 0xAE, 0xAE, 0x39, 0xED, 0xFF, 0xC5,
+        0xC0, 0x8F, 0x08, 0xC8, 0x90, 0x77, 0x30, 0x83, 0x27, 0xEB, 0x0A, 0x34, 0xC7, 0x5B, 0x5A, 0x2A,
+        0xF4, 0x3B, 0x2B, 0xEB, 0x87, 0x57, 0x73, 0x33, 0xFF, 0xEC, 0xBE, 0xD4, 0x62, 0xAE, 0xFC, 0xC5,
+        0xF9, 0xFC, 0x08, 0x91, 0x9D, 0xE2, 0x43, 0xFF, 0xFF, 0xD1, 0x9B, 0x62, 0xE0,
+    ]
+
+    @Test func decodesLongDistanceMatchCorrectly() {
+        let expected = Self.expectedPlaintext()
+        let decoded = LZMADecoder.decode(input: Self.lzmaLongDistanceStream,
+                                         outputLength: expected.count)
+        #expect(decoded == expected)
+    }
+
+    @Test func decodedSecondBlockMatchesFirst() {
+        // The bug corrupted only the copied (matched) region; assert both halves
+        // are byte-identical so a future regression can't pass on the literals alone.
+        let expected = Self.expectedPlaintext()
+        let decoded = LZMADecoder.decode(input: Self.lzmaLongDistanceStream,
+                                         outputLength: expected.count)
+        #expect(decoded?.prefix(256) == decoded?.suffix(256))
+    }
+}
